requirements.txt
# Temel gereksinimler
python-dotenv>=1.0.0
pyttsx3>=2.90
plyer>=2.0.0
keyboard>=0.13.5
pygetwindow>=0.0.9
screeninfo>=0.8
psutil>=5.8.0
schedule>=1.1.0
cryptography>=41.0.0
pycaw>=20181226
screen-brightness-control>=0.9.0
comtypes>=1.1.10
ollama-python>=0.1.0
pathlib>=1.0.1
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
colorama>=0.4.6
PyQt6>=6.4.0
httpx>=0.24.0
requests>=2.31.0
pywin32>=305
pyautogui>=0.9.54
sounddevice>=0.4.6
numpy>=1.21.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
jsonschema>=4.17.3
pycryptodome>=3.19.0

# Ses işleme
gTTS>=2.5.0
edge-tts>=6.1.9
simpleaudio>=1.0.4
playsound>=1.3.0
soundfile>=0.12.1 